Fixed the Endian issue-v1

+++ src/main/java/trafficgenerator/TrafficGenerator.java
@@ -46,7 +46,7 @@ public class TrafficGenerator implements Runnable {
         JpcapSender sender = senderInterface.getSender();
 
 
-        for (int i = 0; i < numberOfPackets; i++) {
+        for (long i = 0; i < numberOfPackets; i++) {
             Packet packet = new Packet();
 
             byte[] mac = ByteOperation.getMacHeader(senderInterface.getSourceMac());
